data: Move undo snapshot restore
The undo data layer separates stack refs and worktree capture from
checkpoint orchestration.

Checkpoint restoration still sits beside command-level undo and redo flow,
so the orchestration module carries blob loading, tree traversal, ref
replay, and worktree writes.

This commit continues separating undo storage by moving checkpoint JSON
reading, tree prefix restoration, ref restoration, worktree restoration,
and intent-to-add replay into data.undo_restore.

The next commit will guard the restore boundary so checkpoint orchestration
cannot grow those helpers again.
