+++ src/git_stage_batch/data/undo.py
@@ -1,21 +1,15 @@
-"""Undo checkpoint storage and restoration."""
+"""Undo checkpoint orchestration."""
 
 from __future__ import annotations
 
 import json
 import os
 import shutil
-import stat
 import tempfile
 from contextlib import contextmanager
 from pathlib import Path
 from typing import Any, Iterator
 
-from ..core.buffer import (
-    write_buffer_to_path,
-    write_buffer_to_working_tree_path,
-)
-from ..utils.repository_buffers import load_git_blob_as_buffer
 from .undo_refs import (
     SESSION_REDO_STACK_REF,
     SESSION_UNDO_STACK_REF,
@@ -25,16 +19,14 @@ from .undo_refs import (
     current_undo_commit,
     list_restorable_refs,
 )
+from . import undo_restore as _undo_restore
 from . import undo_worktree as _undo_worktree
 from ..exceptions import CommandError
 from ..i18n import _
-from ..utils.file_io import read_file_paths_file
 from ..utils.git import (
     create_git_blob,
     GitIndexEntryUpdate,
     get_git_repository_root_path,
-    git_add_paths,
-    git_checkout_detached,
     git_commit_tree,
     git_read_tree,
     git_update_index_entries,
@@ -44,7 +36,6 @@ from ..utils.git import (
     update_git_refs,
 )
 from ..utils.paths import (
-    get_auto_added_files_file_path,
     get_batches_directory_path,
     get_session_directory_path,
     get_state_directory_path,
@@ -74,16 +65,6 @@ def _add_blob_to_index(env: dict[str, str], path: str, data: bytes, mode: str = 
         ],
         env=env,
     )
-
-
-def _restore_file_mode(path: Path, mode: str) -> None:
-    if mode == "120000":
-        return
-    current_mode = path.stat().st_mode
-    if mode == "100755":
-        path.chmod(current_mode | stat.S_IXUSR)
-    else:
-        path.chmod(current_mode & ~stat.S_IXUSR & ~stat.S_IXGRP & ~stat.S_IXOTH)
 
 
 def _add_directory_to_index(env: dict[str, str], *, source_dir: Path, tree_prefix: str) -> None:
@@ -189,7 +170,7 @@ def finalize_pending_checkpoint() -> None:
         return
 
     try:
-        manifest = _read_json_from_commit(checkpoint, "manifest.json")
+        manifest = _undo_restore.read_json_from_commit(checkpoint, "manifest.json")
     except CommandError:
         return
 
@@ -212,123 +193,6 @@ def finalize_pending_checkpoint() -> None:
         message=f"Undo checkpoint: {manifest.get('operation', 'operation')}",
     )
     update_git_refs(updates=[(SESSION_UNDO_STACK_REF, checkpoint_commit)])
-
-
-def _read_json_blob(blob_sha: str) -> dict[str, Any]:
-    with load_git_blob_as_buffer(blob_sha) as buffer:
-        return json.loads(buffer.to_bytes().decode("utf-8"))
-
-
-def _write_blob_to_path(blob_sha: str, target_path: Path) -> None:
-    with load_git_blob_as_buffer(blob_sha) as buffer:
-        write_buffer_to_path(target_path, buffer)
-
-
-def _write_blob_to_worktree_path(
-    blob_sha: str,
-    target_path: Path,
-    *,
-    mode: str,
-) -> None:
-    with load_git_blob_as_buffer(blob_sha) as buffer:
-        write_buffer_to_working_tree_path(target_path, buffer, mode=mode)
-
-
-def _tree_entries(commit: str, prefix: str) -> list[tuple[str, str, str]]:
-    result = run_git_command(
-        ["ls-tree", "-r", "-z", commit, prefix],
-        check=False,
-        text_output=False,
-        requires_index_lock=False,
-    )
-    if result.returncode != 0 or not result.stdout:
-        return []
-
-    entries = []
-    for record in result.stdout.rstrip(b"\0").split(b"\0"):
-        if not record:
-            continue
-        meta, path_bytes = record.split(b"\t", 1)
-        mode, object_type, object_sha = meta.decode("ascii").split()
-        if object_type != "blob":
-            continue
-        entries.append((mode, object_sha, path_bytes.decode("utf-8", errors="surrogateescape")))
-    return entries
-
-
-def _read_json_from_commit(commit: str, path: str) -> dict[str, Any]:
-    entries = _tree_entries(commit, path)
-    if not entries:
-        raise CommandError(_("Undo checkpoint is missing {path}").format(path=path))
-    _mode, blob_sha, _entry_path = entries[0]
-    return _read_json_blob(blob_sha)
-
-
-def _restore_tree_prefix(commit: str, *, prefix: str, target_dir: Path) -> None:
-    if target_dir.exists():
-        shutil.rmtree(target_dir)
-    entries = _tree_entries(commit, prefix)
-    if not entries:
-        return
-    target_dir.mkdir(parents=True, exist_ok=True)
-    for mode, blob_sha, tree_path in entries:
-        relative_path = Path(tree_path).relative_to(prefix)
-        target_path = target_dir / relative_path
-        target_path.parent.mkdir(parents=True, exist_ok=True)
-        _write_blob_to_path(blob_sha, target_path)
-        _restore_file_mode(target_path, mode)
-
-
-def _restore_refs(saved_refs: dict[str, str]) -> None:
-    current_refs = list_restorable_refs()
-    update_git_refs(
-        updates=sorted(saved_refs.items()),
-        deletes=sorted(ref_name for ref_name in current_refs if ref_name not in saved_refs),
-    )
-
-
-def _restore_worktree(commit: str, manifest: dict[str, Any]) -> None:
-    repo_root = get_git_repository_root_path()
-    worktree_blobs = {
-        Path(tree_path).relative_to("worktree").as_posix(): (mode, blob_sha)
-        for mode, blob_sha, tree_path in _tree_entries(commit, "worktree")
-    }
-
-    for entry in manifest.get("worktree_paths", []):
-        file_path = entry["path"]
-        target_path = repo_root / file_path
-        if entry.get("kind") == "gitlink":
-            worktree_oid = entry.get("worktree_oid")
-            if entry.get("exists", False) and worktree_oid:
-                result = git_checkout_detached(worktree_oid, cwd=file_path, check=False)
-                if result.returncode != 0:
-                    raise CommandError(
-                        _("Failed to restore submodule pointer for {file}: {error}").format(
-                            file=file_path,
-                            error=result.stderr,
-                        )
-                    )
-            continue
-        if entry.get("kind") == "embedded-repo":
-            continue
-        if not entry.get("exists", False):
-            if os.path.lexists(target_path):
-                target_path.unlink()
-            continue
-
-        blob_info = worktree_blobs.get(file_path)
-        if blob_info is None:
-            continue
-        mode, blob_sha = blob_info
-        _write_blob_to_worktree_path(blob_sha, target_path, mode=mode)
-
-
-def _restore_intent_to_add_entries() -> None:
-    repo_root = get_git_repository_root_path()
-    for file_path in read_file_paths_file(get_auto_added_files_file_path()):
-        full_path = repo_root / file_path
-        if os.path.lexists(full_path):
-            git_add_paths([file_path], intent_to_add=True, check=False)
 
 
 def _worktree_state_by_path(entries: list[dict[str, Any]]) -> dict[str, dict[str, Any]]:
@@ -485,7 +349,7 @@ def undo_last_checkpoint(*, force: bool = False) -> str:
     if checkpoint is None:
         raise CommandError(_("Nothing to undo."))
 
-    manifest = _read_json_from_commit(checkpoint, "manifest.json")
+    manifest = _undo_restore.read_json_from_commit(checkpoint, "manifest.json")
     conflicts = _detect_conflicts(manifest)
     if conflicts and not force:
         preview = ", ".join(conflicts[:5])
@@ -511,16 +375,24 @@ def undo_last_checkpoint(*, force: bool = False) -> str:
         if live_batches_dir.exists():
             shutil.copytree(live_batches_dir, redo_batches_dir, dirs_exist_ok=True)
 
-        _restore_tree_prefix(checkpoint, prefix="session", target_dir=live_session_dir)
-        _restore_tree_prefix(checkpoint, prefix="batches", target_dir=live_batches_dir)
-        _restore_refs(manifest.get("refs", {}))
+        _undo_restore.restore_tree_prefix(
+            checkpoint,
+            prefix="session",
+            target_dir=live_session_dir,
+        )
+        _undo_restore.restore_tree_prefix(
+            checkpoint,
+            prefix="batches",
+            target_dir=live_batches_dir,
+        )
+        _undo_restore.restore_refs(manifest.get("refs", {}))
 
         index_tree = manifest.get("index_tree")
         if index_tree:
             git_read_tree(index_tree)
 
-        _restore_worktree(checkpoint, manifest)
-        _restore_intent_to_add_entries()
+        _undo_restore.restore_worktree(checkpoint, manifest)
+        _undo_restore.restore_intent_to_add_entries()
 
         after_undo = _snapshot_current_state(redo_paths)
 
@@ -553,7 +425,7 @@ def redo_last_checkpoint(*, force: bool = False) -> str:
     if redo_node is None:
         raise CommandError(_("Nothing to redo."))
 
-    manifest = _read_json_from_commit(redo_node, "manifest.json")
+    manifest = _undo_restore.read_json_from_commit(redo_node, "manifest.json")
     conflicts = _detect_redo_conflicts(manifest)
     if conflicts and not force:
         preview = ", ".join(conflicts[:5])
@@ -564,16 +436,24 @@ def redo_last_checkpoint(*, force: bool = False) -> str:
               "Run 'git-stage-batch redo --force' to overwrite those changes.").format(items=preview)
         )
 
-    _restore_tree_prefix(redo_node, prefix="session", target_dir=get_session_directory_path())
-    _restore_tree_prefix(redo_node, prefix="batches", target_dir=get_batches_directory_path())
-    _restore_refs(manifest.get("refs", {}))
+    _undo_restore.restore_tree_prefix(
+        redo_node,
+        prefix="session",
+        target_dir=get_session_directory_path(),
+    )
+    _undo_restore.restore_tree_prefix(
+        redo_node,
+        prefix="batches",
+        target_dir=get_batches_directory_path(),
+    )
+    _undo_restore.restore_refs(manifest.get("refs", {}))
 
     index_tree = manifest.get("index_tree")
     if index_tree:
         git_read_tree(index_tree)
 
-    _restore_worktree(redo_node, manifest)
-    _restore_intent_to_add_entries()
+    _undo_restore.restore_worktree(redo_node, manifest)
+    _undo_restore.restore_intent_to_add_entries()
 
     undo_checkpoint = manifest.get("undo_checkpoint")
     if undo_checkpoint:
